Restructured API, closes #3
udev was turned into a packge named pyudev.  The namespace of this package
is controlled by apipkg, thus the public interface of pyudev is identical to
udev, but interally code is split into different modules.  The required
apipkg module was added to the source tree and listed as install
requirement.  qudev was moved into the pyudev package as pyudev.pyqt4.
Tests and documentation were updated accordingly.