Work around missing contextlib.nested on Python 3.2 in testsuite
The testsuite must run from Pyhton 2.6 to Python 3.2.  However, Python 3.2
doesn't provide contextlib.nested anymore.  The new syntax of "with", which
makes "nested" superfluous, was introduced in Python 2.7, and isn't available
on Python 2.6.  Thus the only way to nest context managers in all versions is
to provide a "nested()" implementation for Python 3.2 ourselves.  This is done
by conftest now, the "portable" "nested()" is made available in the "pytest"
namespace.