Fix Device.children test failure
Apparently we can't rely on Device.parent == parent for any parent
returned by .match_parent(). udev seems to have a somewhat mysterious
understanding of parent-child relationsships...

Hence, the requirement on this test is loosened, so that the device is
only required to be an ancestors, and the oddness is documented for
Device.children.