Replace use of non-standard fixtures with use of hypothesis.
This gives fewer distracting skip messages, because the skipped tests
get filtered out during input selection by hypothesis.

This results in fewer tests being assembled and fewer skipped.

Runtime is more variable, because of hypothesis database. Should expect it
to be consistently longer on Travis CI by a factor as large as 2.

Some of the @given() annotations could be usefully abstracted once it is
certain that they are correct.

Signed-off-by: mulhern <amulhern@redhat.com>