Separate the way of getting attributes into two separate methods.
Document the methods fully.

Do not test. The methods behave so differently on most machines that there
is no property worth testing.

It is possible to write a version of _get_attributes_sysfs() that does
not make use of _is_attribute_file(), but instead, __contains__. This is
a much more correct method, since it returns a subset of the attributes
returned by _get_attributes_libudev(). Unfortunately, it also actually
reads the attribute values, which is a change in semantics which may
be undesirable.

Signed-off-by: mulhern <amulhern@redhat.com>