Make Attributes class no longer a subset of Mapping.
Closes: #104

Resolves: rhbz#1267584

A general requirement of a mapping is that there is a value for every key,
and that there is not a value for things that are not keys, but this is not
the case for attributes.

This removes some methods and changes the semantics of some other methods.

* __contains__ is no longer defined.

* __getitem__ is no longer defined. There remains a get() method with
the usual semantics for a dict.

* asstring() used to assume that the attribute always had some value, now it
also handles the None case.

It no longer possesses the following concrete methods of Mapping:
__eq__
__ne__
items()
iteritems()
iterkeys()
itervalues()
keys()
values()

Add a property, available_attributes which yields the list of "available"
attributes that libudev provides.

Signed-off-by: mulhern <amulhern@redhat.com>