Make Attributes class no longer a subset of Mapping.
Closes: #104

Resolves: rhbz#1267584

A general requirement of a mapping is that there is a value for every key,
but this is not really the case for attributes.

This removes some methods and changes the semantics of some other methods.

* __contains__ used to return True only if there was a value corresponding
to the key. Now, it returns True if the key is what libudev calls available.
It no longer reads a value for the key, which is probably closer to the
correct semantics.

* __getitem__ used to raise a KeyError for any key that had no value,
regardless of whether the key was an available attribute or not.
Now it raises a KeyError only when the attribute is not among the available
attributes, and returns None when the attribute is among the available
attributes but has no value. This is fairer than raising a KeyError even
if the key is available, which is a bit counter-intuitive.

* asstring() used to assume that the attribute always had some value, now it
also handles the None case.

It no longer possesses the following concrete methods of Mapping:
__eq__
__ne__
get()
items()
iteritems()
iterkeys()
itervalues()
keys()
values()

The only behavior that would have given consistancy with a Mapping interface
would be to only return keys that have a corresponding value. It is,
however, not recommended to read attributes without knowing exactly what
they correspond to, as reading an attribute can have an effect.

Signed-off-by: mulhern <amulhern@redhat.com>