Stop treating Attributes as any sort of mapping from keys to values.
It turns out that the libudev methods used to get the keys and those that
look up the values have a very uncertain relationship to each other, and
the best that you can say is that the intersection between the set of keys
and the strings for which libudev will succesfully look up a value is
almost certainly non-empty.

This change forces getting rid of the __getitem__ method in the Attributes
class, because if it is still around then the 'in' operator will pass
int values to it. Change the method to one called 'lookup' to emphasize
a little more that this is not simply an access to readily available data
using a known useful key.

Signed-off-by: mulhern <amulhern@redhat.com>