Bug fix: allows defaults in all Attributes.as* methods.
Since Attributes can not inherit the Mapping interface, these methods can
not usefully raise a KeyError.

On the other hand, it is reasonable to expect that they will return values
in their designated type.

The string "None" is a bad choice to represent a null result,
because it is a non-empty string.
The empty string is a somewhat better choice, so that is what
asstring() now returns if lookup is unsuccesful. Previously it returned
"None".  Prior to commit 609a08c3cb4baf09135be344bc8913cdbbe5d053, the
method would raise a KeyError.

This does not affect the behavior of asint(), or asbool(), which will raise
a ValueError on "None" or "".

Going forward, this allows setting a default which might be meaningful for
asint() or asbool() methods.

Of course, using asstring(), asbool(), asint(), prevents the client code
from distinguishing a value that was not found from a value that was found
or is the empty string. In general, these are methods best avoided.

Signed-off-by: mulhern <amulhern@redhat.com>