Use DM_MULTIPATH_TIMESTAMP instead of DM_MULTIPATH_DEVICE_NAME.
Two not multipathed devices that point to the same disk will still behave
in the same way. If they have been multipathed in the past,
DM_MULTIPATH_TIMESTAMP might still be hanging around to warn that the
test should be skipped. This check is not actually correct, because there
is no way to precisely identify when the test should be skipped from
pyudev's internals, but it does reduce the number of false positives.

Signed-off-by: mulhern <amulhern@redhat.com>