Remove .github/workflows/weekly file
This file's sole purpose was to run tests weekly on the future Fedora in
order to check for breaking changes. Unfortunately, when these tests
consistently pass, they are forgotten. This can lead to them not being
updated for the new release. They are proving a waste of space, and
generally can be dropped.

Signed-off-by: mulhern <amulhern@redhat.com>